fix: repositioned toast and moved hooks to subfolder
