+++ src/pages/word/semantics/index.tsx
@@ -12,6 +12,8 @@ import { ROUTES } from 'constants/routes';
 import { createSemanticDraggables } from 'utils/words';
 import Meta from 'components/meta';
 import metaTags from 'constants/meta';
+import { showToastMessage } from 'utils';
+import { processWords, semanticsOnDrag, updateSemanticWordsData } from './hooks';
 
 export default function Semantics() {
   const { t: text } = useTranslation();
@@ -42,108 +44,27 @@ export default function Semantics() {
 
   const onDragEnd = (result: DropResult) => {
     const { source, destination } = result;
-    const newWords = Array.from(words);
-    let newSynonyms = Array.from(synonyms);
-    let newAntonyms = Array.from(antonyms);
-
-    let newData = Array.from(words);
-    switch (source.droppableId) {
-      case synonymsText.toLowerCase():
-        newData = newSynonyms;
-        break;
-      case antonymsText.toLowerCase():
-        newData = newAntonyms;
-        break;
-      default:
-        break;
-    }
-    const [foundItem] = newData.splice(source.index, 1);
-
-    if (!destination) return;
-    if (source.droppableId !== destination.droppableId) {
-      const sourceId = source.droppableId;
-      const destinationId = destination.droppableId;
-      if (sourceId === text('ALL_WORDS')) {
-        const newWordLists = newWords.filter((word) => word.id !== foundItem.id);
-        switch (destinationId) {
-          case synonymsText.toLowerCase():
-            setSynonyms([...synonyms, foundItem]);
-            setWords(newWordLists);
-            break;
-          case antonymsText.toLowerCase():
-            setAntonyms([...antonyms, foundItem]);
-            setWords(newWordLists);
-            // check if word is already in antonyms for current word
-            // the following code doesn't allow users to make mistakes
-            // if (antonyms.find((antonym) => antonym.id === foundItem.id)) {
-            //   return;
-            // } else {
-            //   if (currentWord.antonyms?.includes(Number(foundItem.id))) {
-            //   }
-            // }
-            break;
-          default:
-            break;
-        }
-      } else {
-        switch (destinationId) {
-          case text('ALL_WORDS'):
-            if (sourceId === synonymsText.toLowerCase()) {
-              newSynonyms = synonyms.filter((synonym) => synonym.id !== foundItem.id);
-              setSynonyms(newSynonyms);
-              setWords([...newWords, foundItem]);
-            } else if (sourceId === antonymsText.toLowerCase()) {
-              newAntonyms = antonyms.filter((antonym) => antonym.id !== foundItem.id);
-              setAntonyms(newAntonyms);
-              setWords([...newWords, foundItem]);
-            }
-            break;
-          case synonymsText.toLowerCase():
-            if (sourceId === antonymsText.toLowerCase()) {
-              newAntonyms = antonyms.filter((antonym) => antonym.id !== foundItem.id);
-              setAntonyms(newAntonyms);
-              setSynonyms([...newSynonyms, foundItem]);
-            }
-            break;
-          case antonymsText.toLowerCase():
-            if (sourceId === synonymsText.toLowerCase()) {
-              newSynonyms = synonyms.filter((synonym) => synonym.id !== foundItem.id);
-              setSynonyms(newSynonyms);
-              setAntonyms([...newAntonyms, foundItem]);
-            }
-            break;
-          default:
-            break;
-        }
-      }
-    } else {
-      newData.splice(destination.index, 0, foundItem);
-      if (source.droppableId === synonymsText.toLowerCase()) setSynonyms(newData);
-      else if (source.droppableId === antonymsText.toLowerCase()) setAntonyms(newData);
-      else setWords(newData);
-    }
+    semanticsOnDrag(
+      source,
+      destination,
+      text,
+      words,
+      synonyms,
+      antonyms,
+      setWords,
+      setSynonyms,
+      setAntonyms,
+    );
   };
 
   useEffect(() => {
-    const updateWordsData = () => {
-      const wordsData = wordData.filter((word) => word.id && word.id !== Number(wordId));
-      // map through synonyms and antonyms and set its type to 'synonym' or 'antonym'
-      const newWordsData: WordData[] = wordsData.map((word) => {
-        if (currentWord.synonyms?.includes(Number(word.id))) {
-          return { ...word, type: 'synonym' };
-        }
-        if (currentWord.antonyms?.includes(Number(word.id))) {
-          return { ...word, type: 'antonym' };
-        }
-        return word;
-      });
-      // save all to words state
-      setWords(newWordsData);
-    };
-
-    // Call the function unconditionally
-    updateWordsData();
-  }, [wordId, currentWord]);
+    updateSemanticWordsData(
+      wordId,
+      currentWord,
+      wordData,
+      setWords,
+    );
+  }, []);
 
   useEffect(() => {
     const synonymFound = currentWord.synonyms ? currentWord.synonyms.map((synonymId) => synonyms.some(synonym => synonym.id === synonymId)) : [];
@@ -157,16 +78,10 @@ export default function Semantics() {
     }
   }, [currentWord, synonyms, antonyms]);
 
-  const showToastMessage = () => {
-    toast.success(text('SEMANTICS_PRAISE'), {
-      position: toast.POSITION.BOTTOM_RIGHT,
-      closeOnClick: true,
-    });
-  };
 
   useEffect(() => {
     if (showConfetti) {
-      showToastMessage();
+      showToastMessage(text('SEMANTICS_PRAISE'), toast.POSITION.TOP_CENTER, true);
       setTimeout(() => {
         setShowConfetti(false);
       }, 8000);
@@ -174,32 +89,30 @@ export default function Semantics() {
   }, [showConfetti]);
 
   useEffect(() => {
-    const processWords = (wordList: WordData[], type: keyof WordData) => {
-      wordList.forEach((word) => {
-        if (!(currentWord[type] as (number | string)[]).includes(Number(word.id))) {
-          jumpBoxRef.current?.classList.remove('bg-green-500');
-          jumpBoxRef.current.classList.add('jump-box');
-          jumpBoxRef.current.classList.add('bg-red-500');
-          setTimeout(() => {
-            jumpBoxRef.current?.classList.remove('jump-box');
-
-            if (type === 'synonyms') {
-              setSynonyms(synonyms.filter((synonym) => synonym.id !== Number(word.id)));
-            } else if (type === 'antonyms') {
-              setAntonyms(antonyms.filter((antonym) => antonym.id !== Number(word.id)));
-            }
-
-            setWords([...words, word]);
-          }, 1000);
-        } else {
-          jumpBoxRef.current.classList.add('bg-green-500');
-          jumpBoxRef.current.classList.add('z-0');
-        }
-      });
-    };
-
-    processWords(synonyms, 'synonyms');
-    processWords(antonyms, 'antonyms');
+    processWords(
+      synonyms,
+      text('SYNONYMS'),
+      jumpBoxRef,
+      currentWord,
+      words,
+      synonyms,
+      antonyms,
+      setWords,
+      setSynonyms,
+      setAntonyms,
+    );
+    processWords(
+      antonyms,
+      text('ANTONYMS'),
+      jumpBoxRef,
+      currentWord,
+      words,
+      synonyms,
+      antonyms,
+      setWords,
+      setSynonyms,
+      setAntonyms,
+    );
   }, [synonyms, antonyms, words, currentWord]);
 
   if (!currentWord.word) {
