+++ src/utils/index.tsx
@@ -1,0 +1,8 @@
+import { ToastPosition, toast } from 'react-toastify';
+
+export const showToastMessage = (message: string, position: ToastPosition, closeOnClick: boolean) => {
+  toast.success(message, {
+    position: position,
+    closeOnClick,
+  });
+};
